Check for 'create_project' from either the 'flow' or 'microscopy' core technologies.
The original 'microscopy' core technology was separated into 'flow' and 'microscopy', and this required the 'create_project' ingestion service to be replicated. Since in openBIS you cannot register two plug-ins with the same name, the two clones were renamed into 'flow_create_project' and 'micr_create_project'. In the Annotation Tool we can use either one; but we have to check for both since we do not know which core technologies are enabled.